+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Version 0.2.0
 -------------
 
 - Implement the `**` operator.
+- Added :class:`gf256.GF256LT`.
 
 
 Version 0.1.0